+++ chatio-frontend/dev-dist/sw.js
@@ -82,7 +82,7 @@ define(['./workbox-e0e326ff'], (function (workbox) { 'use strict';
     "revision": "3ca0b8505b4bec776b69afdba2768812"
   }, {
     "url": "index.html",
-    "revision": "0.7p8mvtodrkg"
+    "revision": "0.n764pm37838"
   }], {});
   workbox.cleanupOutdatedCaches();
   workbox.registerRoute(new workbox.NavigationRoute(workbox.createHandlerBoundToURL("index.html"), {
