+++ chatio-frontend/package.json
@@ -14,6 +14,7 @@
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-redux": "^9.1.2",
+    "react-router-dom": "^6.27.0",
     "zustand": "^5.0.1"
   },
   "devDependencies": {
