+++ chatio-frontend/src/App.tsx
@@ -1,43 +1,12 @@
-import { useEffect, useState } from 'react';
+import AppLayout from "./layouts/AppLayout";
+import AppRoutes from "./routes";
 
 function App() {
-  const [deferredPrompt, setDeferredPrompt] = useState<Event | null>(null);
-
-  useEffect(() => {
-    const handleBeforeInstallPrompt = (e: Event) => {
-      e.preventDefault();
-      setDeferredPrompt(e);
-    };
-
-    window.addEventListener('beforeinstallprompt', handleBeforeInstallPrompt);
-
-    return () => {
-      window.removeEventListener('beforeinstallprompt', handleBeforeInstallPrompt);
-    };
-  }, []);
-
-  const handleInstallClick = () => {
-    if (deferredPrompt) {
-      (deferredPrompt as any).prompt();
-      (deferredPrompt as any).userChoice.then((choiceResult: any) => {
-        if (choiceResult.outcome === 'accepted') {
-          console.log('User accepted the install prompt');
-        } else {
-          console.log('User dismissed the install prompt');
-        }
-        setDeferredPrompt(null);
-      });
-    }
-  };
-
   return (
     <>
-      <div className='bg-slate-400 w-full h-[800px] text-blue-950'>
-        <h1>Welcome to Chatio!</h1>
-        <button id="install-button" onClick={handleInstallClick}>
-          Install App
-        </button>
-      </div>
+      <AppLayout>
+        <AppRoutes />
+      </AppLayout>
     </>
   );
 }
