+++ chatio-frontend/src/routes/index.tsx
@@ -1,0 +1,17 @@
+// src/routes/index.tsx
+import React from 'react';
+import { BrowserRouter as Router, Route, Routes, Navigate } from 'react-router-dom';
+import Base from '../pages/Base';
+
+const AppRoutes: React.FC = () => {
+  return (
+    <Router>
+      <Routes>
+        <Route path="/" element={<Base />} />
+        <Route path="*" element={<Navigate to="/" />} />
+      </Routes>
+    </Router>
+  );
+};
+
+export default AppRoutes;
